Merge master and hack on highligh current waypoint.
It still does not work, especially if you keep collapsing/expanding
the schematic view as the journey progresses.
My guess is that since Android reuses convertView's we cannot do
it by working with a particular convertView. We must therefore check
on each View's update, whether it is the current waypoint/segment
and set its colour accordingly.